+++ src/popup/popup.js
@@ -8,8 +8,11 @@ import './components/icon';
 import './components/console-command.js';
 import './components/releases-list.js';
 
-import { Release } from '../app/release.js';
-import { getCurrentTab, getExtensionManifest } from '../utils/chrome';
+import {
+  chromeSendMessageToCurrentTab,
+  getCurrentTab,
+  getExtensionManifest
+} from '../utils/chrome';
 import { loadDiscogsGenres } from '../discogs/modules/genres.js';
 import { loadKeywordMapping } from '../bandcamp/modules/mapping.js';
 import config from '../config.js';
@@ -26,7 +29,8 @@ import { bytesToSize } from '../utils/utils';
 import { setupConsole, setupConsoleRelease } from './console.js';
 import { isValidBandcampURL } from '../bandcamp/modules/html.js';
 import { isValidDiscogsReleaseEditUrl } from '../discogs/app/utils.js';
-import { log, logInfo } from '../utils/console';
+import { logInfo } from '../utils/console';
+import { createReleaseFromSchema } from '../utils/schema';
 
 const btnDashboardTab = document.getElementById('dashboard-tab');
 const btnReleaseTab = document.getElementById('release-tab');
@@ -40,47 +44,22 @@ const btnDiscogsSearchArtist = document.getElementById('discogsSearchArtist');
 const tabReleases = document.getElementById('releases');
 
 async function proceedBandcampData() {
-  logInfo('Proceed bandcamp data');
+  logInfo('Proceed Bandcamp data');
 
-  getCurrentTab().then((tab) => {
-    chrome.tabs.sendMessage(tab.id, { type: 'getBandcampData' }, (response) => {
-      if (
-        response === null ||
-        typeof response === 'undefined' ||
-        Object.keys(response).length === 0 ||
-        typeof response.data === 'undefined'
-      ) {
-        showBandcampDataNotFoundWarning();
-        return;
-      }
-
-      processBandcampResponse(response);
-    });
-  });
+  chromeSendMessageToCurrentTab(
+    { type: 'B2D_BC_DATA' },
+    processBandcampResponse,
+    showBandcampDataNotFoundWarning
+  );
 }
 
 async function proceedDiscogsEditPageData() {
-  logInfo('Proceed discogs edit page data');
+  logInfo('Proceed Discogs edit page data');
 
-  getCurrentTab().then((tab) => {
-    log('Sending message `getDiscogsEditPageData` to current tab', tab);
-    chrome.tabs.sendMessage(
-      tab.id,
-      { type: 'getDiscogsEditPageData' },
-      (response) => {
-        if (
-          response === null ||
-          typeof response === 'undefined' ||
-          Object.keys(response).length === 0 ||
-          typeof response.data === 'undefined'
-        ) {
-          return;
-        }
-
-        processDiscogsDraftPageResponse(response);
-      }
-    );
-  });
+  chromeSendMessageToCurrentTab(
+    { type: 'B2D_DISCOGS_EDIT_PAGE_DATA' },
+    processDiscogsDraftPageResponse
+  );
 }
 
 function showBandcampDataNotFoundWarning() {
@@ -108,33 +87,34 @@ function showDashboard() {
 }
 
 function processBandcampResponse(response) {
-  const isRelease = response.type === 'release';
-  const isList = response.type === 'list';
+  const isPageAlbum = response.type === 'TYPE_PAGE_ALBUM';
+  const isPageMusic = response.type === 'TYPE_PAGE_MUSIC';
 
-  if (!isRelease && !isList) {
+  if (!isPageAlbum && !isPageMusic) {
     // todo: Show error?
     return;
   }
 
   loadDiscogsGenres(config.genres_url).then(() => {
     loadKeywordMapping(config.keyword_mapping_url).then((keywordsMapping) => {
-      if (isRelease) {
-        processBandcampReleaseData(response.data, keywordsMapping);
-      } else {
-        processBandcampReleasesData(response);
+      if (isPageAlbum) {
+        processBandcampPageAlbumResponse(response, keywordsMapping);
+      } else if (isPageMusic) {
+        processBandcampPageMusicResponse(response);
       }
     });
   });
 }
 
-function processBandcampReleaseData(data, keywordsMapping) {
+function processBandcampPageAlbumResponse(response, keywordsMapping) {
   hide(btnReleasesTab);
   show(btnReleaseTab);
   click(btnReleaseTab);
 
   try {
-    const release = Release.fromStorageObject(data);
-    setupConsoleRelease(release, keywordsMapping);
+    const schema = response.schema;
+    const release = createReleaseFromSchema(schema);
+    setupConsoleRelease(release, keywordsMapping, schema);
     setupReleaseTab(
       document.getElementById('release'),
       release,
@@ -147,7 +127,7 @@ function processBandcampReleaseData(data, keywordsMapping) {
   }
 }
 
-function processBandcampReleasesData(response) {
+function processBandcampPageMusicResponse(response) {
   hide(btnReleaseTab);
   show(btnReleasesTab);
   click(btnReleasesTab);