+++ src/utils/chrome.ts
@@ -1,9 +1,24 @@
-import { Metadata } from 'src/discogs/app/metadata.js';
+import { logInfo } from './console';
 
 export async function getCurrentTab(): Promise<chrome.tabs.Tab> {
   const queryOptions = { active: true, currentWindow: true };
   const tabs = await chrome.tabs.query(queryOptions);
   return tabs[0];
+}
+
+export async function getCurrentTabUrl(): Promise<string | undefined> {
+  try {
+    const currentTab = await getCurrentTab();
+    if (currentTab.url) {
+      return currentTab.url;
+    } else {
+      console.warn("The current tab doesn't have a URL.");
+      return undefined;
+    }
+  } catch (error) {
+    console.error('Failed to get the current tab URL:', error);
+    return undefined;
+  }
 }
 
 export function getExtensionManifest(): chrome.runtime.Manifest {
@@ -36,18 +51,45 @@ export function getExtensionUrl(path: string): string {
   return chrome.runtime.getURL(path);
 }
 
-interface MetadataMessage {
+interface B2DTabMessage {
   type: string;
-  metadata: Metadata;
 }
 
+type ResponseCallback = (response: any) => void;
+
 export function chromeSendMessageToCurrentTab(
-  message: MetadataMessage,
-  responseCallback?: (response: any) => void
+  message: B2DTabMessage,
+  onValidResponseCallback?: ResponseCallback,
+  onInvalidResponseCallback?: ResponseCallback,
+  responseCallback?: ResponseCallback
 ): void {
+  logInfo('Send message to current tab', message.type, message);
   getCurrentTab().then((tab: chrome.tabs.Tab | undefined) => {
     if (!tab?.id) return;
-    if (responseCallback) {
+
+    if (onValidResponseCallback) {
+      chrome.tabs.sendMessage(tab.id, message, (response) => {
+        logInfo(
+          'Received message from the current tab',
+          message.type,
+          response
+        );
+
+        if (
+          response === null ||
+          typeof response === 'undefined' ||
+          Object.keys(response).length === 0
+        ) {
+          if (onInvalidResponseCallback) {
+            onInvalidResponseCallback(response);
+          }
+
+          return;
+        }
+
+        onValidResponseCallback(response);
+      });
+    } else if (responseCallback) {
       chrome.tabs.sendMessage(tab.id, message, responseCallback);
     } else {
       chrome.tabs.sendMessage(tab.id, message);